+++ lwpreflight/gcp/permission.go
@@ -8,7 +8,12 @@ func CheckPermissions(p *Preflight) error {
 	for _, integrationType := range p.integrationTypes {
 		p.verboseWriter.Write(fmt.Sprintf("Checking permissions for %s", integrationType))
 
-		for _, permission := range RequiredPermissions[integrationType] {
+		requiredPermissions := RequiredPermissions
+		if p.orgID != "" {
+			requiredPermissions = RequiredPermissionsForOrg
+		}
+
+		for _, permission := range requiredPermissions[integrationType] {
 			if !p.permissions[permission] {
 				p.errors[integrationType] = append(
 					p.errors[integrationType],