+++ RELEASE_NOTES.md
@@ -1,9 +1,10 @@
 # Release Notes
-Another day, another release. These are the release notes for the version `v2.3.2`.
+Another day, another release. These are the release notes for the version `v2.5.0`.
 
+## Features
+* feat(lwpreflight): support org level checks (#1744) (Pengyuan Zhao)([573f48c1](https://github.com/lacework/go-sdk/commit/573f48c160cb9b7f892e2cee94575188c8156a1f))
 ## Other Changes
-* chore: check permissions with wildcard (#1738) (lokeshv-fortinet)([4a3c9893](https://github.com/lacework/go-sdk/commit/4a3c989383793c125f33ea8eedef66f376088a49))
-* ci: version bump to v2.3.2-dev (Lacework)([9654bcda](https://github.com/lacework/go-sdk/commit/9654bcda183de4426e19fd44db99702b0f0999a9))
+* ci: version bump to v2.4.1-dev (Lacework)([3018ef02](https://github.com/lacework/go-sdk/commit/3018ef024291bdea12063bbe326d5fd8e4bf7773))
 
 ## :whale: [Docker Image](https://hub.docker.com/r/lacework/lacework-cli)
 ```